// frontend/src/client/types.gen.ts
export type UpdateStaffCommand = {
    firstName?: string;
    lastName?: string;
    email?: string;
    active?: boolean;
    specialization?: string;
    pwz?: string;
    department?: string;
    position?: string;
    workSchedule?: string;
};